+++ internal/models/coupon.go
@@ -7,7 +7,7 @@ type Coupon struct {
 	ProductID int       `json:"product_id"`
 	UserID    int       `json:"user_id"`
 	Code      string    `json:"code"`
-	Rate      float64   `json:"rate"`
+	Rate      float32   `json:"rate"`
 	ExpireAt  time.Time `json:"expire_at"`
 	Active    bool      `json:"active"`
 	CreatedAt time.Time `json:"-"`
