+++ internal/models/history.go
@@ -7,7 +7,7 @@ type History struct {
 	ProductID int       `json:"product_id"`
 	UserID    int       `json:"user_id"`
 	Quantity  int       `json:"quantity"`
-	Discount  float64   `json:"discount"`
+	Discount  float32   `json:"discount"`
 	CreatedAt time.Time `json:"-"`
 	UpdatedAt time.Time `json:"-"`
 }
@@ -17,7 +17,7 @@ type HistoryGet struct {
 	ProductName string    `json:"product_name"`
 	Quantity    int       `json:"quantity"`
 	UserName    string    `json:"first_name"`
-	Discount    int       `json:"discount"`
+	Discount    float32   `json:"discount"`
 	CreatedAt   time.Time `json:"-"`
 	UpdatedAt   time.Time `json:"-"`
 }
